Refactor EnvironmentGateway into modular projects
Updated the solution to include new projects: EnvironmentGatewayApi, EnvironmentGatewayDomain, EnvironmentGatewayApplication, EnvironmentGatewayInfrastructure, and SharedKernel. Removed the previous EnvironmentGateway project reference and adjusted the Global section accordingly.

Added a project reference to EnvironmentGatewayApi in DevServer.AppHost.csproj. Modified Program.cs to integrate the new API and removed outdated gateway configurations.

Updated the Dockerfile to support the new project structure and created new HTTP files for EnvironmentGatewayApi.

Introduced WeatherForecast and WeatherForecastController classes in the new API. Updated appsettings files for logging configurations.

These changes enhance the modularity and maintainability of the EnvironmentGateway functionality.
