+++ src/DevServer.AppHost/DevServer.AppHost.csproj
@@ -18,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\EnvironmentGateway\EnvironmentGateway.csproj" />
+    <ProjectReference Include="..\EnvironmentGateway\EnvironmentGatewayApi\EnvironmentGatewayApi.csproj" />
     <ProjectReference Include="..\WeatherForecastApi\WeatherForecastApi.csproj" />
   </ItemGroup>
