+++ src/EnvironmentGateway/EnvironmentGatewayApi/Dockerfile
@@ -12,19 +12,19 @@ EXPOSE 8081
 FROM mcr.microsoft.com/dotnet/sdk:9.0 AS build
 ARG BUILD_CONFIGURATION=Release
 WORKDIR /src
-COPY ["src/EnvironmentGateway/EnvironmentGateway.csproj", "src/EnvironmentGateway/"]
-RUN dotnet restore "./src/EnvironmentGateway/EnvironmentGateway.csproj"
+COPY ["src/EnvironmentGateway/EnvironmentGatewayApi/EnvironmentGatewayApi.csproj", "src/EnvironmentGateway/EnvironmentGatewayApi/"]
+RUN dotnet restore "./src/EnvironmentGateway/EnvironmentGatewayApi/EnvironmentGatewayApi.csproj"
 COPY . .
-WORKDIR "/src/src/EnvironmentGateway"
-RUN dotnet build "./EnvironmentGateway.csproj" -c $BUILD_CONFIGURATION -o /app/build
+WORKDIR "/src/src/EnvironmentGateway/EnvironmentGatewayApi"
+RUN dotnet build "./EnvironmentGatewayApi.csproj" -c $BUILD_CONFIGURATION -o /app/build
 
 # This stage is used to publish the service project to be copied to the final stage
 FROM build AS publish
 ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "./EnvironmentGateway.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "./EnvironmentGatewayApi.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
 
 # This stage is used in production or when running from VS in regular mode (Default when not using the Debug configuration)
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "EnvironmentGateway.dll"]
+ENTRYPOINT ["dotnet", "EnvironmentGatewayApi.dll"]
