+++ src/EnvironmentGateway/EnvironmentGatewayApi/appsettings.json
