--- src/EnvironmentGateway/InMemoryConfig.cs
@@ -1,9 +1,0 @@
-﻿using Yarp.ReverseProxy.Configuration;
-
-namespace EnvironmentGateway;
-
-public class InMemoryConfig
-{
-    public IReadOnlyList<RouteConfig> Routes { get; set; } = new List<RouteConfig>();
-    public IReadOnlyList<ClusterConfig> Clusters { get; set; } = new List<ClusterConfig>();
-}
