--- src/EnvironmentGateway/Models/EnvironmentGateway.cs
@@ -1,3 +1,0 @@
-﻿namespace EnvironmentGateway.Models;
-
-public record EnvironmentGateway(string GatewayName);
