--- src/EnvironmentGateway/appsettings.Development.json
@@ -1,27 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ReverseProxy": {
-    "Routes": {
-      "route1": {
-        "ClusterId": "cluster1",
-        "Match": {
-          "Path": "{**catch-all}"
-        }
-      }
-    },
-    "Clusters": {
-      "cluster1": {
-        "Destinations": {
-          "destination1": {
-            "Address": "https://neuko-know-how.com/"
-          }
-        }
-      }
-    }
-  }
-}
